introduce IGitVersionTool
