--- src/GitVersionCore/Common/IGitVersionCalculator.cs
@@ -1,9 +1,0 @@
-using GitVersion.OutputVariables;
-
-namespace GitVersion
-{
-    public interface IGitVersionCalculator
-    {
-        VersionVariables CalculateVersionVariables();
-    }
-}
